backend/requirements.txt
beautifulsoup4~=4.9.3
django-extensions~=3.1.3
django_cron~=0.5.1
djangorestframework~=3.12.4
django~=3.2.4
gunicorn~=20.1.0
pillow~=8.2.0
psycopg2-binary~=2.9.1
python-magic~=0.4.24
PyYAML~=5.4.1
rawkit~=0.6.0
requests~=2.25.1
stegano~=0.9.8
uritemplate~=3.0.1
